move out of legacy directory
