+++ gateway-js/src/IntrospectAndCompose/__tests__/tsconfig.json
